refactor(ui): extract health/render/health-overview.ts — risk scoring, cards, and recommendations
Move renderHealthOverview and its risk-scoring + recommendation
pipeline out of health.ts into health/render/health-overview.ts. The
module owns the single source of truth for the Health tab's "Overall
health" status. The barrel re-exports renderHealthOverview so the
existing call sites stay on the same import path.
